refactor: split trades.rs into trades/ directory
Move trades.rs to trades/mod.rs and extract
get_trades_for_transaction into trades/get_by_tx.rs
for better code organization.
